requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
jinja2==3.1.4
python-multipart==0.0.12
pydantic==2.10.0
api-analytics
python-dotenv
pyyaml==6.0.1